+++ src/main/java/com/example/masterplanbbe/common/exception/GlobalException.java
@@ -1,0 +1,32 @@
+package com.example.masterplanbbe.common.exception;
+
+public abstract class GlobalException extends BaseException {
+    protected GlobalException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    public static class NotFoundException extends GlobalException {
+        public NotFoundException(ErrorCode errorCode) {
+            super(errorCode);
+        }
+    }
+
+    public static class InternalServerException extends GlobalException {
+        public InternalServerException(ErrorCode errorCode) {
+            super(errorCode);
+        }
+    }
+
+    public static class IllegalStateException extends GlobalException {
+        public IllegalStateException(ErrorCode errorCode) {
+            super(errorCode);
+        }
+    }
+
+    public static class BadRequestException extends GlobalException {
+        public BadRequestException(ErrorCode errorCode) {
+            super(errorCode);
+        }
+    }
+
+}
