+++ src/main/java/com/example/masterplanbbe/domain/exam/repository/ExamRepositoryAdapter.java
@@ -16,7 +16,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.function.LongSupplier;
 
-import static com.example.masterplanbbe.common.GlobalException.*;
+import static com.example.masterplanbbe.common.exception.GlobalException.*;
 import static com.example.masterplanbbe.common.exception.ErrorCode.*;
 import static com.example.masterplanbbe.domain.exam.entity.QExam.exam;
 import static com.example.masterplanbbe.domain.examBookmark.entity.QExamBookmark.examBookmark;
