+++ src/main/java/com/example/masterplanbbe/domain/post/service/PostService.java
@@ -1,6 +1,6 @@
 package com.example.masterplanbbe.domain.post.service;
 
-import com.example.masterplanbbe.common.GlobalException;
+import com.example.masterplanbbe.common.exception.GlobalException;
 import com.example.masterplanbbe.common.exception.ErrorCode;
 import com.example.masterplanbbe.domain.comment.dto.CommentResponse;
 import com.example.masterplanbbe.domain.post.dto.PostRequest;
